Move the custom command stuff to its own file.
